@anish749 2019-09-16
I don't remember, but did you also try separating out the read and writes into two different `DoFn`s? Does that move the thread management to Beam?