@labianchin 2019-09-17
> I don't remember, but did you also try separating out the read and writes into two different DoFns? Does that move the thread management to Beam?

I remember trying on earlier versions of DBeam to have two phases: read JDBC and write to Avro. Problem was that Beam was waiting for the read bundle to complete, serialize and then write to Avro, which was very inefficient. If we found a way to "stream" from different `DoFn`s (including the fanout: one jdbc query fans out to millions of records), then we could rely on Beam for that.